rename golden_configs -> golden
